+++ tests/testthat/test-check-functions.R
@@ -245,6 +245,32 @@ test_that("'check_life_colnums' returns correct error with by, groups clash", {
                  "Can't supply `by` when `data` is a grouped data frame.")
 })
 
+test_that("'check_life_colnums' returns correct error when mx, by clash", {
+    empty_colnum <- integer()
+    names(empty_colnum) <- character()
+    expect_error(check_life_colnums(mx_colnum = c(mx = 1L),
+                                    qx_colnum = empty_colnum,
+                                    age_colnum = c(AGE = 3L),
+                                    sex_colnum = empty_colnum,
+                                    ax_colnum = c(AX = 2L),
+                                    by_colnums = c(strata = 7L, mx = 1L),
+                                    groups_colnums = empty_colnum),
+                 "`mx` and `by` use the same variable.")
+})
+
+test_that("'check_life_colnums' returns correct error when sex, groups clash", {
+    empty_colnum <- integer()
+    names(empty_colnum) <- character()
+    expect_error(check_life_colnums(mx_colnum = c(mx = 1L),
+                                    qx_colnum = empty_colnum,
+                                    age_colnum = c(AGE = 3L),
+                                    sex_colnum = empty_colnum,
+                                    ax_colnum = c(AX = 2L),
+                                    by_colnums = empty_colnum,
+                                    groups_colnums = c(strata = 7L, mx = 1L)),
+                 "`mx` and `groups` use the same variable.")
+})
+
 
 ## 'check_lx' --------------------------------------------------------
 
@@ -756,8 +782,7 @@ test_that("'check_sex_not_needed' returns correct error when methods do require 
     expect_error(check_sex_not_needed(methods),
                  "`child` is \"CD\" but no value supplied for `sex`")
 })
-
-    
+ 
 
 ## 'check_standard' -----------------------------------------------------------
 
@@ -923,6 +948,71 @@ test_that("'check_target_ex_to_lifetab_brass' throws expected error when no inde
                  "`target` does not have index variables.")
 })
 
+
+## 'check_tfr_colnums' -------------------------------------------------------
+
+test_that("'check_tfr_colnums' returns TRUE with valid inputs - no groups", {
+    empty_colnum <- integer()
+    names(empty_colnum) <- character()
+    expect_true(check_tfr_colnums(asfr_colnum = c(asfr = 1L),
+                                   age_colnum = c(AGE = 3L),
+                                   sex_colnum = empty_colnum,
+                                   by_colnums = c(region = 5L, time = 4L),
+                                   groups_colnums = empty_colnum))
+})
+
+test_that("'check_tfr_colnums' returns correct error when no asfr", {
+  empty_colnum <- integer()
+  expect_error(check_tfr_colnums(asfr_colnum = empty_colnum,
+                                 age_colnum = c(AGE = 3L),
+                                 sex_colnum = empty_colnum,
+                                 by_colnums = c(region = 5L, time = 4L),
+                                 groups_colnums = empty_colnum),
+               "No value supplied for `asfr`.")
+})
+
+test_that("'check_tfr_colnums' returns correct error when no age", {
+  empty_colnum <- integer()
+  expect_error(check_tfr_colnums(asfr_colnum = c(asfr = 3L),
+                                 age_colnum = empty_colnum,
+                                 sex_colnum = empty_colnum,
+                                 by_colnums = c(region = 5L, time = 4L),
+                                 groups_colnums = empty_colnum),
+               "No value supplied for `age`.")
+})
+
+test_that("'check_tfr_colnums' returns correct error with by, groups clash", {
+  empty_colnum <- integer()
+  names(empty_colnum) <- character()
+  expect_error(check_tfr_colnums(asfr_colnum = c(asfr = 1L),
+                                 age_colnum = c(AGE = 3L),
+                                 sex_colnum = empty_colnum,
+                                 by_colnums = c(strata = 7L),
+                                 groups_colnums = c(region = 5L, time = 4L)),
+               "Can't supply `by` when `data` is a grouped data frame.")
+})
+
+test_that("'check_tfr_colnums' returns correct error when asfr, by clash", {
+    empty_colnum <- integer()
+    names(empty_colnum) <- character()
+    expect_error(check_tfr_colnums(asfr_colnum = c(asfr = 1L),
+                                    age_colnum = c(AGE = 3L),
+                                    sex_colnum = empty_colnum,
+                                    by_colnums = c(strata = 7L, asfr = 1L),
+                                    groups_colnums = empty_colnum),
+                 "`asfr` and `by` use the same variable.")
+})
+
+test_that("'check_tfr_colnums' returns correct error when asfr, groups clash", {
+  empty_colnum <- integer()
+  names(empty_colnum) <- character()
+  expect_error(check_tfr_colnums(asfr_colnum = c(asfr = 1L),
+                                 age_colnum = c(AGE = 3L),
+                                 sex_colnum = empty_colnum,
+                                 by_colnums = empty_colnum,
+                                 groups_colnums = c(strata = 7L, asfr = 1L)),
+               "`asfr` and `groups` use the same variable.")
+})
 
 
 ## 'check_valid_colnum_list' --------------------------------------------------