+++ NEWS.md
@@ -13,6 +13,7 @@
   `n_core` to a value greater than 1 triggers parallel processing.
 * Added `closed` argument to `check_age()`, to check whether the
   oldest age group is closed.
+* Added function `tfr()` for calculating total fertility rates.
 
 
 ## Bug fixes